+++ composer.json
@@ -3,7 +3,7 @@
     "description": "A Zipkin integration for Symfony applications",
     "type": "symfony-bundle",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "symfony/config": "^6.0",
         "symfony/http-kernel": "^6.0",
         "symfony/routing": "^6.0",
